feat: simple  address find function(imcomplete)
