+++ src/entities/element/detail/ApplyReject.tsx
