+++ src/entities/element/detail/Footer.tsx
