+++ src/entities/element/detail/Product.tsx
