+++ src/pages/address/AddressPage.tsx
@@ -1,7 +1,7 @@
 // src/App.tsx
 
 import React from 'react';
-import AddressForm from '@/entities/element/AddressForm';
+import AddressForm from '@/entities/element/address/AddressForm';
 
 const App: React.FC = () => {
   return (
